backend/requirements.txt
asgiref==3.2.10
Django==3.1.2
dj-database-url==0.5.0
python-decouple==3.3
django-crispy-forms==1.9.2
django-localflavor==3.0.1
python-stdnum==1.14
pytz==2020.1
sqlparse==0.4.0
validate-docbr==1.7.0
djangorestframework==3.12.1
psycopg2-binary>=2.8
django-cors-headers
